move SearchForm inside CoursePane
