+++ src/components/RightPane/CoursePane/CourseRenderPane.js
@@ -6,8 +6,8 @@ import noNothing from './static/no_results.png';
 import darkNoNothing from './static/dark-no_results.png';
 import AppStore from '../../../stores/AppStore';
 import RightPaneStore from '../../../stores/RightPaneStore';
-import loadingGif from '../SearchForm/Gifs/loading.gif';
-import darkModeLoadingGif from '../SearchForm/Gifs/dark-loading.gif';
+import loadingGif from './SearchForm/Gifs/loading.gif';
+import darkModeLoadingGif from './SearchForm/Gifs/dark-loading.gif';
 import GeDataFetchProvider from '../SectionTable/GEDataFetchProvider';
 import LazyLoad from 'react-lazyload';
 import { queryWebsoc, queryWebsocMultiple, isDarkMode } from '../../../helpers';
