+++ src/components/RightPane/CoursePane/SearchForm/DeptSearchBar/MobileDeptSelector.js
