+++ src/components/RightPane/CoursePane/SearchForm/Gifs/dark-loading.gif
