+++ src/components/RightPane/CoursePane/SearchForm/Gifs/loading.gif
