+++ src/components/RightPane/CoursePane/SearchForm/SectionCodeSearchBar.js
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import { TextField } from '@material-ui/core';
-import { updateFormValue } from '../../../actions/RightPaneActions';
-import RightPaneStore from '../../../stores/RightPaneStore';
+import { updateFormValue } from '../../../../actions/RightPaneActions';
+import RightPaneStore from '../../../../stores/RightPaneStore';
 
 class SectionCodeSearchBar extends PureComponent {
     state = {
